+++ actix-session/src/cookie.rs
@@ -25,7 +25,7 @@ use actix_web::dev::{ServiceRequest, ServiceResponse};
 use actix_web::http::{header::SET_COOKIE, HeaderValue};
 use actix_web::{Error, HttpMessage, ResponseError};
 use derive_more::{Display, From};
-use futures::future::{ok, FutureExt, LocalBoxFuture, Ready};
+use futures_util::future::{ok, FutureExt, LocalBoxFuture, Ready};
 use serde_json::error::Error as JsonError;
 use time::{Duration, OffsetDateTime};
 
@@ -391,11 +391,9 @@ mod tests {
         let mut app = test::init_service(
             App::new()
                 .wrap(CookieSession::signed(&[0; 32]).secure(false))
-                .service(web::resource("/").to(|ses: Session| {
-                    async move {
-                        let _ = ses.set("counter", 100);
-                        "test"
-                    }
+                .service(web::resource("/").to(|ses: Session| async move {
+                    let _ = ses.set("counter", 100);
+                    "test"
                 })),
         )
         .await;
@@ -405,8 +403,7 @@ mod tests {
         assert!(response
             .response()
             .cookies()
-            .find(|c| c.name() == "actix-session")
-            .is_some());
+            .any(|c| c.name() == "actix-session"));
     }
 
     #[actix_rt::test]
@@ -414,11 +411,9 @@ mod tests {
         let mut app = test::init_service(
             App::new()
                 .wrap(CookieSession::private(&[0; 32]).secure(false))
-                .service(web::resource("/").to(|ses: Session| {
-                    async move {
-                        let _ = ses.set("counter", 100);
-                        "test"
-                    }
+                .service(web::resource("/").to(|ses: Session| async move {
+                    let _ = ses.set("counter", 100);
+                    "test"
                 })),
         )
         .await;
@@ -428,8 +423,7 @@ mod tests {
         assert!(response
             .response()
             .cookies()
-            .find(|c| c.name() == "actix-session")
-            .is_some());
+            .any(|c| c.name() == "actix-session"));
     }
 
     #[actix_rt::test]
@@ -437,11 +431,9 @@ mod tests {
         let mut app = test::init_service(
             App::new()
                 .wrap(CookieSession::signed(&[0; 32]).secure(false))
-                .service(web::resource("/").to(|ses: Session| {
-                    async move {
-                        let _ = ses.set("counter", 100);
-                        "test"
-                    }
+                .service(web::resource("/").to(|ses: Session| async move {
+                    let _ = ses.set("counter", 100);
+                    "test"
                 })),
         )
         .await;
@@ -451,8 +443,7 @@ mod tests {
         assert!(response
             .response()
             .cookies()
-            .find(|c| c.name() == "actix-session")
-            .is_some());
+            .any(|c| c.name() == "actix-session"));
     }
 
     #[actix_rt::test]
@@ -468,17 +459,13 @@ mod tests {
                         .same_site(SameSite::Lax)
                         .max_age(100),
                 )
-                .service(web::resource("/").to(|ses: Session| {
-                    async move {
-                        let _ = ses.set("counter", 100);
-                        "test"
-                    }
+                .service(web::resource("/").to(|ses: Session| async move {
+                    let _ = ses.set("counter", 100);
+                    "test"
                 }))
-                .service(web::resource("/test/").to(|ses: Session| {
-                    async move {
-                        let val: usize = ses.get("counter").unwrap().unwrap();
-                        format!("counter: {}", val)
-                    }
+                .service(web::resource("/test/").to(|ses: Session| async move {
+                    let val: usize = ses.get("counter").unwrap().unwrap();
+                    format!("counter: {}", val)
                 })),
         )
         .await;
@@ -505,11 +492,9 @@ mod tests {
         let mut app = test::init_service(
             App::new()
                 .wrap(CookieSession::signed(&[0; 32]).secure(false).expires_in(60))
-                .service(web::resource("/").to(|ses: Session| {
-                    async move {
-                        let _ = ses.set("counter", 100);
-                        "test"
-                    }
+                .service(web::resource("/").to(|ses: Session| async move {
+                    let _ = ses.set("counter", 100);
+                    "test"
                 }))
                 .service(
                     web::resource("/test/")