+++ actix-identity/CHANGES.md
@@ -1,8 +1,10 @@
 # Changes
 
-## [Unreleased] - 2020-xx-xx
+## [0.3.0-alpha.1] - 2020-03-14
 
-* Update the `time` dependency to 0.2.5
+* Update the `time` dependency to 0.2.7
+* Update the `actix-web` dependency to 3.0.0-alpha.1
+* Minimize `futures` dependency
 
 ## [0.2.1] - 2020-01-10