+++ actix-redis/CHANGES.md
@@ -1,5 +1,23 @@
 # Changes
 
+## [Unreleased]
+
+* Update `actix` to 0.10.0-alpha.2
+
+* Update `actix-session` to 0.4.0-alpha.1
+
+* Update `actix-web` to 3.0.0-alpha.1
+
+* Update `time` to 0.2.9
+
+## [0.8.1] 2020-02-18
+
+* Move `env_logger` dependency to dev-dependencies and update to 0.7
+
+* Update `actix_web` to 2.0.0 from 2.0.0-rc
+
+* Move repository to actix-extras
+
 ## [0.8.0] 2019-12-20
 
 * Release